requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
fastapi>=0.115.0
uvicorn>=0.32.0
akshare>=1.14.0
pydantic>=2.9.0
python-dotenv>=1.0.0
httpx>=0.27.0
